requirements.txt
# ============================================
# Multi-Agent Laptop Store Management System
# ============================================

# Web Framework
Flask==3.1.0
flask-cors==5.0.1

# Database
mysql-connector-python==9.1.0
SQLAlchemy==2.0.36
PyMySQL==1.1.1

# Multi-Agent Framework
crewai==0.86.0
crewai-tools==0.14.0
langchain==0.3.12
langchain-openai==0.2.12
langchain-community==0.3.12

# Web Scraping
selenium==4.27.1
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Scheduling
APScheduler==3.10.4

# Environment & Config
python-dotenv==1.0.1

# Authentication
bcrypt==4.2.1

# Production Server
gunicorn==23.0.0

# Utilities
requests==2.32.3
fake-useragent==1.5.1
retry==0.9.2
